Fix non-termination bug in adaptive iterations loop
After we decreased the number of max iterations that is tried
by default it uncovered a bug in the adaptive iterations discovery:

If number of iterations is at maximum, but it still does not satisfy
minimum time requirement it would endless keep retrying the same number
of iterations over and over.

The change also adds an assert to insure we crash in cases like that
in the future.