Switch to using percentage of median.
This new representation gives a much more effective summary of the performance
and variability of the benchmark. It's also much easier on the eyes too!

Along with this, I switched up the way the stats functions were written (to
instead by computed properties of `Array<Double>`) as that avoided the name
clashes with using similarly named local variables.

Note: while the median is the right measurement, and standard deviation is a useful way
of capturing variation, standard deviation and mean are the statistics that usually
go together.