Add support for lower overhead measurement loop (#37)
* Add BenchmarkState for lower overhead measurements

This change adds support for lower-overhead version of
writing explicit measurement loop within the benchmark
itself:

```swift
benchmark("my bench") { state in
    // set-up, not measured

    state.measure {
        // body of the benchmark, runs #iterations times
    }

    // tear-down, not measured
}
```

By using measure we get less overhead of measurements by around 10%:

```
    $ swift run -c release BenchmarkMinimalExample --min-time 10
    --max-iterations 100000000
    Updating https://github.com/apple/swift-argument-parser
    Resolving https://github.com/apple/swift-argument-parser at 0.0.6
    [4/4] Linking BenchmarkMinimalExample
    running noop... done! (6095.58 ms)
    running noop with state... done! (6215.35 ms)
    running measure noop... done! (5549.44 ms)

    name            time    std         iterations
    ----------------------------------------------
    noop            23.0 ns ±  83.88 %   100000000
    noop with state 23.0 ns ± 1473.15 %  100000000
    measure noop    21.0 ns ±  88.80 %   100000000
```

It is the same precision as if the clock is used directly in a tight
loop without any closures. To improve precision further we would need
to provide a higher precision clock then the one provided by
DispatchTime.

Fixes #29

* Add state.measure test

* Drop flags used for performance debugging

* Make state.measure run the closure once

A little bit more context on why methods such as `state.measure` are now throwing: in ML benchmarks, we want to have more control on how the training/inference loop interacts with benchmark measurements.

For example, an inference loop that uses [swift-api](https://github.com/tensorflow/swift-apis)'s Epochs API is going too look like:

```swift
benchmark("...") {
  for epochBatches in dataset.training {
    for batch in epochBatches {
      let images = batch.data
      try state.measure {
        let _ = model(images)
        LazyTensorBarrier()
      }
    }
  }
}
```

Here, the `state.measure` call is enclosed in two levels of looping: an infinite loop over epochs, and a finite loop over all batches within an epoch. 

Our objective is to record *exactly* as many measurements as was requested by the benchmark runner, and stop immediately after this number. Otherwise, the heuristic for automatic detection of the number iterations will not be able to make progress to converge on a good number of iterations with a few retries. 

We achieve this by supporting arbitrary looping within the benchmark with arbitrary amount of calls to `state.measure` that terminates early whenever we've recorded the requested number of iterations. You might wonder what happens if we don't record enough measurements within a single benchmark run, as in:

```swift
benchmark("...", settings: Iterations(3000)) {
  let images = ...
  for _ in 1...1337 {
    try state.measure {
      let _ = model(images)
      LazyTensorBarrier()
    }
  }
}
```

In this case, the body of the benchmark is going to be rerun multiple times until the total number of measurements reaches the target number. Number of iterations don't have to align, the last benchmark run will be terminated mid-way, once 3000 measurements has been collected.

To make all the pieces fit together, we want to support early termination that can't be misused. This is achieved by using `Termination` error for the purpose of breaking out of multiple layers all the way back to the benchmark runner. It's important to highlight that with new `state`-based API we need to be able to unwind arbitrarily levels deep because `state.measure` call can be invoked from anywhere. For example, a more sophisticated training loop can be factored out into a layer of library code and then the model will be invoked within one of its callbacks. Errors work really well here because language already happens to supports a good and compiler checked mechanism for error propagation!

Another solution to support early termination would be to return a result value that makes user terminate his loop manually. This doesn't work nearly as well because it will turn benchmark code into explicit error handling, similar to the `if err != nil { ... }` pattern in Go. This becomes even more problematic once you have to thread this through library code that is not aware of the termination errors (and it shouldn't be!).

* Add doc comments; drop unused APIs

* Rename Termination to BenchmarkTermination

* Update CMakeLists

* Relax measurement tests to avoid flaky cloud ci failures

* Add another test case

* Add missing allTests entry